fix: restructure wallet hooks
