+++ src/components/Pools/AddLiquidity/index.tsx
@@ -18,8 +18,8 @@ import { Field, resetMintState } from 'src/state/pmint/actions';
 import { useDerivedMintInfo, useMintActionHandlers, useMintState } from 'src/state/pmint/hooks';
 import { SpaceType } from 'src/state/pstake/types';
 import { useIsExpertMode, useUserSlippageTolerance } from 'src/state/puser/hooks';
-import { useCurrencyBalance } from 'src/state/pwallet/hooks/common';
 import { useAddLiquidityHook } from 'src/state/pwallet/hooks';
+import { useCurrencyBalance } from 'src/state/pwallet/hooks/common';
 import { maxAmountSpend } from 'src/utils/maxAmountSpend';
 import { wrappedCurrency } from 'src/utils/wrappedCurrency';
 import ConfirmPoolDrawer from './ConfirmPoolDrawer';
