+++ src/components/SarManageWidget/AddStake/index.tsx
@@ -18,7 +18,7 @@ import { useWalletModalToggle } from 'src/state/papplication/hooks';
 import { useDerivativeSarStakeHook } from 'src/state/psarstake/hooks';
 import { useSarStakeInfo } from 'src/state/psarstake/hooks/evm';
 import { Position } from 'src/state/psarstake/types';
-import { useTokenBalanceHook } from 'src/state/pwallet/multiChainsHooks';
+import { useTokenBalanceHook } from 'src/state/pwallet/hooks';
 import { getBuyUrl } from 'src/utils';
 import { hederaFn } from 'src/utils/hedera';
 import ConfirmDrawer from '../ConfirmDrawer';
